Fix script + update README.md

+++ Dark Smooth GUI/zip_versions.sh
@@ -5,13 +5,22 @@ if [ $# != 1 ]; then
 	exit 1
 fi
 
+if [ -f "pack.mcmeta" ]; then
+	rm "pack.mcmeta"
+fi
+
+rm Dark\ Smooth\ GUI*.zip
+
 # Obtains the MC version using each directory
 VERSIONS=$(ls -d */ | sed 's#/##')
 
-echo "$VERSIONS"
-
 for ver in $VERSIONS; do
-	rm -r assets/ pack.mcmeta
+	if [ -d "assets/" ]; then
+		rm -r assets/ pack.mcmeta
+	fi
 	cp -r $ver/assets $ver/pack.mcmeta ./
-	zip -r "Dark Smoot GUI $1+$ver.zip" assets/ LICENSE.txt pack.png pack.mcmeta
+	zip -r "Dark Smooth GUI $1+$ver.zip" assets/ LICENSE.txt pack.png pack.mcmeta
 done
+
+rm -r "assets/"
+rm "pack.mcmeta"
